Fixed min and max numeric parameter range values

+++ SLC_SM_IAS_Profiles/Views/UIRows/NumberRow.cs
@@ -17,8 +17,8 @@
 		{
 			base.Configure();
 
-			double minimum = Data.Record.ConfigurationParameterValue.NumberOptions.MinRange ?? double.MinValue;
-			double maximum = Data.Record.ConfigurationParameterValue.NumberOptions.MaxRange ?? double.MaxValue;
+			double minimum = Data.Record.ConfigurationParameterValue.NumberOptions.MinRange ?? int.MinValue;
+			double maximum = Data.Record.ConfigurationParameterValue.NumberOptions.MaxRange ?? int.MaxValue;
 			int decimalVal = Convert.ToInt32(Data.Record.ConfigurationParameterValue.NumberOptions.Decimals);
 			double stepSize = Data.Record.ConfigurationParameterValue.NumberOptions.StepSize ?? 1;
 			var value = new Numeric(Data.Record.ConfigurationParameterValue.NumberOptions.DefaultValue ?? 0)
